+++ src/main/java/com/hhplus/concert/ConcertApplication.java
@@ -2,8 +2,12 @@ package com.hhplus.concert;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
+@EnableScheduling
 @SpringBootApplication
+@ConfigurationPropertiesScan
 public class ConcertApplication {
 
 	public static void main(String[] args) {
