+++ src/main/java/com/hhplus/concert/core/domain/concert/ConcertSeat.java
@@ -1,4 +1,4 @@
-package com.hhplus.concert.core.domain.concert.entlty;
+package com.hhplus.concert.core.domain.concert;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -12,25 +12,28 @@ import java.time.LocalDateTime;
 @AllArgsConstructor(access = AccessLevel.PACKAGE)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@Table(name = "PAYMENT")
-public class Payment {
+@Table(name = "CONCERT_SEAT")
+public class ConcertSeat {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "user_id", nullable = false)
-    private Long userId;
+    @Column(name = "concert_schedule_id", nullable = false)
+    private Long concertScheduleId;
 
-    @Column(name = "reservation_id", nullable = false)
-    private Long reservationId;
+    @Column(name = "amount", nullable = false)
+    private Integer amount;
 
-    @Column(name = "price", nullable = false)
-    private Long price;
+    @Column(name = "position", nullable = false)
+    private Integer position;
 
     @Enumerated(EnumType.STRING)
-    @Column(name = "status", nullable = false)
-    private PaymentStatus status;
+    @Column(name = "seat_status", nullable = false)
+    private SeatStatus seatStatus;
+
+    @Column(name = "reserved_until_dt")
+    private LocalDateTime reservedUntilDt;
 
     @Column(name = "created_dt", nullable = false)
     private LocalDateTime createdDt;
