+++ src/main/java/com/hhplus/concert/core/domain/concert/PaymentStatus.java
@@ -1,4 +1,4 @@
-package com.hhplus.concert.core.domain.concert.entlty;
+package com.hhplus.concert.core.domain.concert;
 
 public enum PaymentStatus {
     PROGRESS,
