+++ src/main/java/com/hhplus/concert/core/domain/concert/PaymentType.java
@@ -1,0 +1,6 @@
+package com.hhplus.concert.core.domain.concert;
+
+public enum PaymentType {
+    PAYMENT,
+    REFUND
+}
