--- src/main/java/com/hhplus/concert/core/domain/concert/entlty/PaymentType.java
@@ -1,6 +1,0 @@
-package com.hhplus.concert.core.domain.concert.entlty;
-
-public enum PaymentType {
-    PAYMENT,
-    REFUND
-}
