+++ src/main/java/com/hhplus/concert/core/domain/queue/Queue.java
@@ -1,4 +1,4 @@
-package com.hhplus.concert.core.domain.queue.entlty;
+package com.hhplus.concert.core.domain.queue;
 
 import io.jsonwebtoken.Jwts;
 import jakarta.persistence.*;
