+++ src/main/java/com/hhplus/concert/core/domain/queue/SelectQueueTokenResult.java
@@ -1,0 +1,7 @@
+package com.hhplus.concert.core.domain.queue;
+
+public record SelectQueueTokenResult(
+        long queuePosition,
+        QueueStatus status
+) {
+}
