--- src/main/java/com/hhplus/concert/core/domain/queue/dto/SelectQueueTokenResult.java
@@ -1,9 +1,0 @@
-package com.hhplus.concert.core.domain.queue.dto;
-
-import com.hhplus.concert.core.domain.queue.entlty.QueueStatus;
-
-public record SelectQueueTokenResult(
-        long queuePosition,
-        QueueStatus status
-) {
-}
