+++ src/main/java/com/hhplus/concert/core/domain/user/UserService.java
@@ -1,7 +1,5 @@
-package com.hhplus.concert.core.domain.user.service;
+package com.hhplus.concert.core.domain.user;
 
-import com.hhplus.concert.core.domain.user.entlty.Users;
-import com.hhplus.concert.core.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
