+++ src/main/java/com/hhplus/concert/core/infrastructure/repository/user/repository/UserRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.hhplus.concert.core.infrastructure.repository.user.repository;
 
-import com.hhplus.concert.core.domain.user.entlty.Users;
-import com.hhplus.concert.core.domain.user.repository.UserRepository;
+import com.hhplus.concert.core.domain.user.Users;
+import com.hhplus.concert.core.domain.user.UserRepository;
 import com.hhplus.concert.core.infrastructure.repository.user.persistence.UserJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
