+++ src/main/java/com/hhplus/concert/core/interfaces/v1/queue/QueueController.java
@@ -1,7 +1,7 @@
 package com.hhplus.concert.core.interfaces.v1.queue;
 
-import com.hhplus.concert.core.domain.queue.dto.SelectQueueTokenResult;
-import com.hhplus.concert.core.domain.queue.service.QueueService;
+import com.hhplus.concert.core.domain.queue.SelectQueueTokenResult;
+import com.hhplus.concert.core.domain.queue.QueueService;
 import com.hhplus.concert.core.interfaces.common.CommonRes;
 import com.hhplus.concert.core.interfaces.v1.queue.req.CreateQueueTokenReq;
 import com.hhplus.concert.core.interfaces.v1.queue.res.CreateQueueTokenRes;
